+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- [Fixed `parseAmount` function error when amount is invalid number](https://github.com/ElrondNetwork/dapp-core/pull/519)
 
 ## [[2.2.3](https://github.com/ElrondNetwork/dapp-core/pull/510)] - 2022-12-16
 - [Fixed login on nativeAuth block failed](https://github.com/ElrondNetwork/dapp-core/pull/516)